changed package name to personName and refactored public items for better readability with personName. leading
