+++ pkg/personName/info.go
@@ -1,17 +1,17 @@
 /*
-The pn package provides methods and data types for inspecting Person Name (PN)
+The personName package provides methods and data types for inspecting Person Name (PN)
 Value Representations.
  */
-package pn
+package personName
 
 import (
 	"errors"
 	"regexp"
 )
 
-// Parsed PersonName value information from an element with a "PN" VR. See the "PN"
+// Parsed Info value information from an element with a "PN" VR. See the "PN"
 // entry at: http://dicom.nema.org/dicom/2013/output/chtml/part05/sect_6.2.html
-type PersonName struct {
+type Info struct {
 	// The person's family or last name.
 	FamilyName string
 	// The person's given or first names
@@ -25,7 +25,7 @@ type PersonName struct {
 }
 
 // Returns dicom format PN string: '[Last]^[First]^[Middle]^[Prefix]^[Suffix]'
-func (pn PersonName) String() string {
+func (pn Info) String() string {
 	return pn.FamilyName +
 		"^" + pn.GivenName +
 		"^" + pn.MiddleName +
@@ -38,17 +38,17 @@ var pnRegex = regexp.MustCompile(
 	`(?P<Fist>.*)\^(?P<Last>.*)\^(?P<Middle>.*)\^(?P<Prefix>.*)\^(?P<Suffix>.*)`,
 )
 
-// ErrParsePersonName is returned when attempting to parse PersonName from a string.
+// ErrParsePersonName is returned when attempting to parse Info from a string.
 var ErrParsePersonName = errors.New("person name value does not match Dicom Spec" +
 	" '[Last]^[First]^[Middle]^[Prefix]^[Suffix]'")
 
-// Converts a string from a dicom element with a Value Representation of PN to a
-// parsed PersonName struct.
-func NewPersonNameFromDicom(pnString string) (name PersonName, err error) {
+// FromDicomValueString converts a string from a dicom element with a Value
+// Representation of PN to a parsed Info struct.
+func FromDicomValueString(pnString string) (name Info, err error) {
 	// Run the regex against the name
 	matches := pnRegex.FindStringSubmatch(pnString)
 	if len(matches) == 0 || matches[0] == "" {
-		return PersonName{}, ErrParsePersonName
+		return Info{}, ErrParsePersonName
 	}
 	name.FamilyName = matches[1]
 	name.GivenName = matches[2]
